CRDL-231: Separated OutboundSoapMessage file
